BACK-END/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.3.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6
httpx>=0.25.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
posthog==3.5.0
